Refactor usage error and remove hint for non-usage-related errors
